Remove text that would help with SEO
I've started hunting for links to this FAQ and trying to point them to the new repo (e.g. [this one](http://stackoverflow.com/questions/10122977/what-are-the-best-practices-for-structuring-a-large-meteor-app-with-many-html-te#comment18069856_13039681)).
Problem is, Google searches all point here, and having any real content here will perpetuate the situation. Hence the proposal to just link to the new FAQ here.